moved polyfills out of main js file
